Переписан главный экран с трекерами под архитектуру MVVM.
Исправлен баг во время поиска(после нажатия на кнопку отменить появлялась заглушка вместо трекеров)
