+++ Tracker/Controllers/Trackers/CreateTracker/CreateTrackerViewController.swift
@@ -11,7 +11,6 @@ import SnapKit
 final class CreateTrackerViewController: UIViewController, CreateTrackerViewControllerProtocol {
     
     private let newTrackerView = CreateTrackerView()
-    var presenter: TrackerViewPresenterProtocol?
     var viewController: TrackerViewControllerProtocol?
     
     override func viewDidLoad() {
@@ -42,7 +41,6 @@ final class CreateTrackerViewController: UIViewController, CreateTrackerViewCont
     func switchToTrackerVC() {
         dismiss(animated: true)
         viewController?.setupTrackersFromDatePicker()
-        viewController?.reloadCollectionView()
     }
 }
