+++ Tracker/Controllers/Trackers/CreateTracker/Protocols/CreateTrackerViewControllerProtocol.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 protocol CreateTrackerViewControllerProtocol: AnyObject {
-    var presenter: TrackerViewPresenterProtocol? { get set }
     var viewController: TrackerViewControllerProtocol? { get }
     func switchToTrackerVC()
 }
