+++ Tracker/Controllers/Trackers/Protocols/TrackerViewControllerProtocol.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 protocol TrackerViewControllerProtocol: AnyObject {
-    var presenter: TrackerViewPresenterProtocol? { get set }
     func reloadCollectionView()
     func setupTrackersFromDatePicker()
 }
