--- Tracker/Controllers/Trackers/Protocols/TrackerViewPresenterProtocol.swift
@@ -1,20 +1,0 @@
-//
-//  TrackerViewPresenterProtocol.swift
-//  Tracker
-//
-//  Created by Dmitry Medvedev on 29.05.2023.
-//
-
-import Foundation
-
-protocol TrackerViewPresenterProtocol: AnyObject {
-    var currentDate: Date? { get set }
-    func checkDate() -> Bool
-    func getVisibleCategories() -> [TrackerCategory]
-    func getCompletedCategories() -> [TrackerRecord]
-    func fetchCompletedCategoriesFromStore()
-    func filterTrackers(text: String?)
-    func deleteTracker(id: UUID)
-    func addRecord(tracker: TrackerRecord)
-    func deleteRecord(tracker: TrackerRecord)
-}
