+++ Tracker/Controllers/Trackers/Protocols/TrackersViewModelProtocol.swift
@@ -8,5 +8,12 @@
 import Foundation
 
 protocol TrackersViewModelProtocol: AnyObject {
-    
+    func deleteTracker(id: UUID)
+    func fetchCompletedCategoriesFromStore()
+    func getCompletedCategories() -> [TrackerRecord]
+    func addRecord(tracker: TrackerRecord)
+    func deleteRecord(tracker: TrackerRecord)
+    func setupVisibleTrackers()
+    func checkDate() -> Bool
+    func filterTrackers(text: String?)
 }
