+++ Tracker/Services/Protocols/TrackerStoreProtocol.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 protocol TrackerStoreProtocol: AnyObject {
-    var delegate: TrackersDelegate? { get set }
     func addTracker(_ model: Tracker)
     func getTracker(category: String, index: Int) -> Tracker
     func fetchTrackers() -> [TrackerCategory]
