+++ Tracker/TabBarController.swift
@@ -20,9 +20,7 @@ final class TabBarController: UITabBarController {
         appearance.backgroundColor = .ypWhite
         self.tabBar.standardAppearance = appearance
         
-        let trackerViewPresenter = TrackerViewPresenter()
         let trackerViewController = TrackersViewController()
-        trackerViewController.presenter = trackerViewPresenter
         
         let trackerNavigationController = UINavigationController(rootViewController: trackerViewController)
         let statisticNavigationController = UINavigationController(rootViewController: StatisticViewController())
